+++ packages/ragbits-document-search/CHANGELOG.md
@@ -5,6 +5,7 @@
 - add docling document parser (#509)
 - move sources from ragbits-document-search to ragbits-core (#496)
 - fix union types validation in element enricher (#499)
+- disable unstructured analytics due to performance issues (#521)
 
 ## 0.13.0 (2025-04-02)